~ move specs regarding cache to its own files
