+++ spec/record/has_many_spec.rb
@@ -45,73 +45,7 @@ describe LHS::Record do
     end
   end
 
-  # TODO move
-  context 'clearing cache when including associations' do
-    before do
-      class Place < LHS::Record
-        endpoint 'https://datastore/places/{id}', followlocation: true, headers: { 'Prefer' => 'redirect-strategy=redirect-over-not-found' }
-        has_many :available_assets
-      end
-
-      class AvailableAsset < LHS::Record
-      end
-
-      stub_request(:get, "http://datastore/places/#{place_id}/available-assets?limit=100")
-        .to_return(body: {
-          total: available_assets.size,
-          items: available_assets
-        }.to_json)
-    end
-
-    let(:place_id) { SecureRandom.urlsafe_base64 }
-
-    let(:place_hash) do
-      {
-        href: "https://datastore/places/#{place_id}",
-        id: place_id,
-        available_assets: { href: "http://datastore/places/#{place_id}/available-assets?offset=0&limit=10" }
-      }
-    end
-
-    let(:available_asset_hash) do
-      { asset_code: 'OPENING_HOURS' }
-    end
-
-    let(:available_assets) { [available_asset_hash] }
-
-    it 'clears the cache when using find' do
-      stub_request(:get, place_hash[:href])
-        .to_return(body: place_hash.to_json)
-      place = Place
-        .options(auth: { bearer: 'XYZ' })
-        .includes_all(:available_assets)
-        .find(place_id)
-      expect(place.available_assets.first).to be_a(AvailableAsset)
-    end
-
-    it 'clears the cache when using where' do
-      stub_request(:get, place_hash[:href])
-        .to_return(body: place_hash.to_json)
-      place = Place
-        .options(auth: { bearer: 'XYZ' })
-        .includes_all(:available_assets)
-        .where(id: place_id)
-      expect(place.available_assets.first).to be_a(AvailableAsset)
-    end
-
-    it 'clears the cache when using find_by' do
-      stub_request(:get, "https://datastore/places/#{place_id}?limit=1")
-        .to_return(body: place_hash.to_json)
-      place = Place
-        .options(auth: { bearer: 'XYZ' })
-        .includes_all(:available_assets)
-        .find_by(id: place_id)
-      expect(place.available_assets.first).to be_a(AvailableAsset)
-    end
-  end
-
   context 'custom class_name' do
-
     before do
       module Uberall
         class Location < LHS::Record
